+++ .gitignore
@@ -1,5 +1,2 @@
 /target/
 /.idea/
-/src/test/java/jp/go/aist/dggs/CheckDistortion.java
-/src/main/java/org/jengineering/sjmply/
-/src/main/java/org/smurn/jply/
